pip_requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anykeystore==0.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.4
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
cryptacular==1.6.2
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
datasets==2.0.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepspeed==0.6.0
dill==0.3.4
docstring-parser==0.13
fairscale==0.4.6
filelock==3.6.0
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1647963865961/work
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
google-auth==2.6.2
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
hjson==3.0.2
huggingface-hub==0.4.0
hupper==1.10.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
jsonargparse==4.4.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1647351495879/work
librosa==0.9.1
lightning-flash @ git+https://github.com/PyTorchLightning/lightning-flash.git@90014499886d533265f7f6f70747b3d70b05f0db
llvmlite==0.38.0
Markdown==3.3.6
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1639359646028/work
multidict==6.0.2
multiprocess==0.70.12.2
munkres==1.1.4
ninja==1.10.2.3
nltk==3.7
numba==0.55.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1640083068831/work
nvidia-ml-py==11.450.51
nvitop==0.5.3
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.1
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1602493092985/work
plaster==1.0
plaster-pastedeploy==0.7
pooch==1.6.0
protobuf==3.19.4
psutil==5.9.0
py-cpuinfo==8.0.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1642753572664/work
pyramid==2.0
pyramid-mailer==0.15.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python3-openid==3.2.0
pytorch-lightning==1.5.10
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
repoze.sendmail==4.4.1
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
resampy==0.2.2
responses==0.18.0
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
sacremoses==0.0.49
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1640464150285/work
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six @ file:///tmp/build/80754af9/six_1623709665295/work
SoundFile==0.10.3.post1
SQLAlchemy==1.4.32
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.7.3
torchvision==0.12.0
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
transaction==3.0.1
transformers==4.17.0
translationstring==1.4
typing_extensions==4.1.1
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1640031013105/work
urllib3==1.26.7
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.7
Werkzeug==2.0.3
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xxhash==3.0.0
yarl==1.7.2
zipp==3.7.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
